Unify configuration file structure
This changes a lot of small details in the configuration file in an
attempt to unify its structure. These are some main guidelines used for
this refactoring:
    - Specify that changes require restart consistently
    - Unify the specification of available field values
    - Provide clear distinction between description title and body

Besides these guidelines used to unify minor details in the
configuration file, the section on key configuration has been completely
reworked in an attempt to reduce the amount of text used. This should
make it possible to understand what's going on without having to read
any text.

The notice that modifiers are not supported has been removed from the
mouse binding documentation.