Fix mesa rendering outside window borders on wayland
The mesa workaround has lead to some issues with
rendering on Wayland.

To resolve this problem, the mesa workaround has been
restructured in a way which still allows clearing the screen
before rendering without killing performance with the mesa
driver. The performance is identical to the master brach
and there have been no recorded regressions.