Fix selection of empty lines
When selecting multiple lines in Alacritty, there was an issue with
empty lines not being copied. This behavior has been chanaged so empty
lines should be correctly copied now.

When copying content which ends with an empty line, Alacritty would copy
an additional empty line.

This has been resolved by only adding empty lines when the empty line
was not in the last selected line.