Fix selection with right and middle mouse
Since there was no check for the button state for semantic and line
selection, it was possible to trigger selection using the middle and
right mouse button. This has been resolved by explicitly checking for
the pressed button before starting these selections.

To further ensure that double and triple clicks are only triggered when
the same button is pressed multiple times, the last pressed button is
stored.

This fixes #1915.