Fix selection copying with mouse outside of window
There was an issue caused by 35efb4619c4b7a77b3c30de763856bc4441e236e
which would lead to the current selection not getting copied to the
clipboard if the mouse was released outside of the window.

Instead of aborting any press/release actions when the cursor is not
inside of Alacritty, the handling is now delayed until actual usage.