Add new icon to macOS binary
This adds the new Alacritty icon to the macOS binary available in the
github releases. On sizes 128x128 and up the image with scanlines is
included.

The ci deployment process should now also offer the svg file for
download so the .desktop file can be used without having to look through
the repos for the icon.